refactor: ♻️ supabase 요청 함수 정리
